Modifier masks are NSUInteger

+++ Cocoa/KeyboardShortcutPrivateAPIs.h
@@ -4,10 +4,10 @@
 @interface NSKeyboardShortcut : NSObject <NSCopying>
 
 + (id)shortcutWithPreferencesEncoding:(NSString *)encoding;
-+ (id)shortcutWithKeyEquivalent:(NSString *)key_equivalent modifierMask:(unsigned long long)mask;
-- (id)initWithKeyEquivalent:(NSString *)key_equivalent modifierMask:(unsigned long long)mask;
++ (id)shortcutWithKeyEquivalent:(NSString *)key_equivalent modifierMask:(NSUInteger)mask;
+- (id)initWithKeyEquivalent:(NSString *)key_equivalent modifierMask:(NSUInteger)mask;
 
-@property(readonly) unsigned long long modifierMask;
+@property(readonly) NSUInteger modifierMask;
 @property(readonly) NSString *keyEquivalent;
 @property(readonly) NSString *preferencesEncoding;
 @property(readonly) NSString *localizedModifierMaskDisplayName;
